refactor: Improvements with the location of the DTOs
