+++ src/main/java/com/app/shopin/modules/security/dto/auth/JwtDTO.java
@@ -1,4 +1,4 @@
-package com.app.shopin.modules.security.dto;
+package com.app.shopin.modules.security.dto.auth;
 
 public class JwtDTO {
     private String token;
