+++ src/main/java/com/app/shopin/modules/security/dto/auth/SetNewPasswordDTO.java
@@ -1,4 +1,4 @@
-package com.app.shopin.modules.security.dto;
+package com.app.shopin.modules.security.dto.auth;
 
 public record SetNewPasswordDTO(String validationToken, String newPassword) {}
