+++ src/main/java/com/app/shopin/modules/security/dto/oauth2/CompleteRegistrationDTO.java
@@ -1,4 +1,4 @@
-package com.app.shopin.modules.security.dto;
+package com.app.shopin.modules.security.dto.oauth2;
 
 import jakarta.validation.constraints.NotBlank;
