+++ src/main/java/com/app/shopin/modules/security/dto/oauth2/OAuth2TempInfo.java
@@ -1,4 +1,4 @@
-package com.app.shopin.modules.security.dto;
+package com.app.shopin.modules.security.dto.oauth2;
 
 public record OAuth2TempInfo(String provider,
                              String providerUserId,
