+++ src/main/java/com/app/shopin/modules/security/dto/twofactor/CodeConfirmationDTO.java
@@ -1,4 +1,4 @@
-package com.app.shopin.modules.security.dto;
+package com.app.shopin.modules.security.dto.twofactor;
 
 import jakarta.validation.constraints.NotBlank;
