+++ src/main/java/com/app/shopin/modules/security/dto/twofactor/SetPreferredTwoFactorDTO.java
@@ -1,4 +1,4 @@
-package com.app.shopin.modules.security.dto;
+package com.app.shopin.modules.security.dto.twofactor;
 
 import com.app.shopin.modules.security.enums.TwoFactorMethod;
 import jakarta.validation.constraints.NotNull;
