+++ src/main/java/com/app/shopin/modules/security/dto/twofactor/SetupTwoFactorDTO.java
@@ -1,4 +1,4 @@
-package com.app.shopin.modules.security.dto;
+package com.app.shopin.modules.security.dto.twofactor;
 
 public record SetupTwoFactorDTO(String manualEntryKey, String qrCodeImageUri) {}
