+++ src/main/java/com/app/shopin/modules/security/jwt/OAuth2LoginSuccessHandler.java
@@ -1,6 +1,6 @@
 package com.app.shopin.modules.security.jwt;
 
-import com.app.shopin.modules.security.dto.OAuth2TempInfo;
+import com.app.shopin.modules.security.dto.oauth2.OAuth2TempInfo;
 import com.app.shopin.modules.security.entity.PrincipalUser;
 import com.app.shopin.modules.user.entity.User;
 import jakarta.servlet.ServletException;
